rename: MemePost service, repository 폴더 생성
