+++ src/main/java/com/findmymeme/memepost/repository/MemePostRepository.java
@@ -1,5 +1,6 @@
-package com.findmymeme.memepost.domain;
+package com.findmymeme.memepost.repository;
 
+import com.findmymeme.memepost.domain.MemePost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemePostRepository extends JpaRepository<MemePost, Long> {
